feat: 質問詳細ページのコンポーネントを修正
