+++ frontend/src/pages/questions/detail/[id].tsx
@@ -6,7 +6,9 @@ import { useRecoilValue } from "recoil";
 
 import { signInUserState } from "@/store/Auth/auth";
 import PostPreview from "@/components/Questions/PostPreview";
+import AnswerList from "@/components/Answer";
 import Layout from "@/components/Layouts/layout";
+import AnswerForm from "@/components/Answer/AnswerForm";
 
 interface Questions {
   id: string;
@@ -29,7 +31,6 @@ interface Answers {
   createdAt: Timestamp;
   updatedAt: Timestamp;
 }
-
 
 const QuestionDetail = () => {
   const router = useRouter();
@@ -90,34 +91,6 @@ const QuestionDetail = () => {
     }
   }, [id]);
 
-  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    const db = getFirestore();
-    try {
-      const newAnswerRef = await addDoc(collection(db, "questions", id as string, "answers"), {
-        qId: id,
-        uId: uid, // replace with actual user ID
-        mId: question?.mId,
-        content: newAnswer,
-        createdAt: Timestamp.now(),
-        updatedAt: Timestamp.now(),
-      });
-      setNewAnswer("");
-
-      const newAnswerData = {
-        ...newAnswerRef,
-        qId: id,
-        uId: uid,
-        mId: question?.mId,
-        content: newAnswer,
-        createdAt: Timestamp.now(),
-        updatedAt: Timestamp.now(),
-      } as Answers;
-      setAnswers([...answers, newAnswerData]);
-    } catch (error) {
-      console.error(error);
-    }
-  };
 
   return (
     <Layout>
@@ -127,27 +100,8 @@ const QuestionDetail = () => {
             <>
               <Title>{question.title}</Title>
               <PostPreview markdown={question.content} />
-              <AnswerList>
-                <h2>回答一覧</h2>
-                {answers.map((answer) => (
-                  <Answer key={answer.id}>
-                    <AnswerContent>{answer.content}</AnswerContent>
-                    <AnswerCreatedAt>
-                      {new Date(
-                        answer.createdAt.seconds * 1000 + answer.createdAt.nanoseconds / 1000000
-                      ).toLocaleString()}
-                    </AnswerCreatedAt>
-                  </Answer>
-                ))}
-              </AnswerList>
-              <NewAnswerForm onSubmit={handleSubmit}>
-                <NewAnswerInput
-                  value={newAnswer}
-                  onChange={(e) => setNewAnswer(e.target.value)}
-                  placeholder="Enter your answer here"
-                />
-                <NewAnswerButton type="submit">Post</NewAnswerButton>
-              </NewAnswerForm>
+              <AnswerList answers={answers} />
+              <AnswerForm id={id as string} setAnswers={setAnswers} answers={answers} mId={question.mId} />
             </>
           )}
         </Wrapper>
@@ -161,11 +115,11 @@ export default QuestionDetail;
 const Container = styled.div`
   display: flex;
   justify-content: center;
-  height: 100vh;
+  margin-bottom: 4rem;
 `;
 
 const Wrapper = styled.div`
-  width: 80%;
+  width: 100%;
   max-width: 800px;
 `;
 
@@ -173,49 +127,4 @@ const Title = styled.h1`
   font-size: 2rem;
   font-weight: bold;
   margin-bottom: 1rem;
-`;
-
-const Answer = styled.li`
-  margin-bottom: 1.5rem;
-`;
-
-const NewAnswerButton = styled.button`
-  background-color: #4caf50;
-  color: white;
-  border: none;
-  padding: 0.5rem 1rem;
-  border-radius: 5px;
-  cursor: pointer;
-`;
-
-const NewAnswerForm = styled.form`
-  display: flex;
-  flex-direction: column;
-  margin-top: 2rem;
-`;
-
-const NewAnswerInput = styled.textarea`
-  height: 10rem;
-  margin-bottom: 1rem;
-  padding: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-`;
-
-const AnswerList = styled.ul`
-  margin-top: 3rem;
-  padding: 0;
-  list-style: none;
-  margin-left: 1rem;
-`;
-
-const AnswerContent = styled.div`
-  padding: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-`;
-
-const AnswerCreatedAt = styled.span`
-  font-size: 0.8rem;
-  color: #666;
 `;
